+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.83.0
 
 * `[fmtutil/panel]` Added `TOP_LINE` option
+* `[strutil]` Added method `LenVisual`
 * `[fmtutil/panel]` Improved panel rendering when `BOTTOM_LINE` option is used
 * `[fmtutil/panel]` Improved panel rendering when `label` is empty
 * `[fmtutil/panel]` Added limit for minimal panel size (_256 symbols_)